# ggml: cache sin/cos for RoPE