On master the ggml RoPE kernel currently iterates over all rows and computes a pair of sin/cos values per 2 columns. However, these sin/cos values do not depend on the row. As a consequence the RoPE kernel does a lot of redundant computation. This PR simply caches the sin/cos values so that they are only computed once by each thread. This makes evaluation a bit faster if there is specialized hardware for matrix multiplication (e.g. a GPU), especially if the CPU has a low number of cores. 7b q4_0 performance changes as follows on my system with a Ryzen 3700X and an RTX 3090:

| GPU layers | CUDA | Threads | Test  | t/s master | t/s ggml-rope-cache | Speedup |
|------------|------|---------|-------|------------|---------------------|---------|
|          0 | No   |       4 | pp512 |      13.33 |               13.28 |    1.00 |
|          0 | No   |       8 | pp512 |      25.63 |               25.62 |    1.00 |
|          0 | Yes  |       1 | pp512 |     175.54 |              199.13 |    1.13 |
|          0 | Yes  |       1 | tg128 |       3.17 |                3.19 |    1.01 |
|          0 | Yes  |       2 | pp512 |     242.87 |              264.30 |    1.09 |
|          0 | Yes  |       2 | tg128 |       6.09 |                6.13 |    1.01 |
|          0 | Yes  |       4 | pp512 |     299.25 |              313.31 |    1.05 |
|          0 | Yes  |       4 | tg128 |      10.46 |               10.46 |    1.00 |
|          0 | Yes  |       8 | pp512 |     321.25 |              333.10 |    1.04 |
|          0 | Yes  |       8 | tg128 |      10.39 |               10.41 |    1.00 |
|         15 | Yes  |       1 | pp512 |     255.77 |              282.63 |    1.11 |
|         15 | Yes  |       1 | tg128 |       5.62 |                5.80 |    1.03 |
|         15 | Yes  |       2 | pp512 |     371.49 |              402.20 |    1.08 |
|         15 | Yes  |       2 | tg128 |      10.81 |               10.85 |    1.00 |
|         15 | Yes  |       4 | pp512 |     477.13 |              503.38 |    1.06 |
|         15 | Yes  |       4 | tg128 |      17.89 |               17.95 |    1.00 |
|         15 | Yes  |       8 | pp512 |     543.09 |              559.56 |    1.03 |
|         15 | Yes  |       8 | tg128 |      17.64 |               17.56 |    1.00 |
|         30 | Yes  |       1 | pp512 |    1694.35 |             1839.85 |    1.09 |
|         30 | Yes  |       1 | tg128 |      29.80 |               30.33 |    1.02 |
|         30 | Yes  |       2 | pp512 |    1979.55 |             2065.12 |    1.04 |
|         30 | Yes  |       2 | tg128 |      47.09 |               47.49 |    1.01 |
|         30 | Yes  |       4 | pp512 |    2157.50 |             2214.33 |    1.03 |
|         30 | Yes  |       4 | tg128 |      63.67 |               64.05 |    1.01 |
|         30 | Yes  |       8 | pp512 |    2230.74 |             2256.93 |    1.01 |
|         30 | Yes  |       8 | tg128 |      60.20 |               60.35 |    1.00 |

For the implementation I used `wdata` to cache the values. I didn't notice any issues but how much space can I assume to be available in `wdata`?