@JohannesGaessler 2024-01-13
It seems the CI tests were failing because `tests/test-rope` did not have `wdata`. I edited `ggml_graph_plan` to consider RoPE for `wsize` which presumably also answers my question about how much `wdata` I can assume to be available.