@ggerganov 2024-01-13
We can extract the cache generation code into a simple static function `ggml_rope_cache_init()` and reuse it for the F16 rope as well