@JohannesGaessler 2024-01-13
If we do that then I think we should also cover the other branches in `ggml_compute_forward_rope` (e.g. glm). I don't have models with which to test any such implementations. Can I assume that `tests/test-rope` will catch any bugs?