@ggerganov 2024-01-13
`tests/test-backend-ops -b CUDA0 -o ROPE` will catch any issues because it will compare the results against the CUDA implementation. It currently covers F32 and F16, standard and NeoX mode. GLM mode is not covered though.