@JohannesGaessler 2024-01-13
In the end I decided to only implement caching for non-glm, non-neox RoPE because previously I had incorrectly assumed that the same sin/cos values would be needed for all three branches.